refactor: move code generation code to dedicated generator.go file
